backend/requirements.txt
# Backend Requirements - FastAPI + LLM + Database
# Generated: 2026-01-10

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9

# Redis Cache
redis>=5.0.1

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.5.3
pydantic-settings>=2.1.0

# OpenAI LLM
openai>=1.10.0

# Ontology Processing
rdflib>=7.0.0

# Rate Limiting
slowapi>=0.1.9

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Monitoring (optional)
prometheus-client>=0.19.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
httpx>=0.26.0
